Build: Copy libmonoboehm-2.0.so.1 to libmonoboehm-2.0.so while building
We need the packaged library to be called libmonoboehm-2.0.so.1 because
that is the DT_SONAME we are linking against.

During linking, we need it to be available under the libmonoboehm-2.0.so
name.  This is usually done by linking:

    ln -s libmonoboehm-2.0.so.1 libmonoboehm-2.0.so

However,

 1. We do not know if the build host supports POSIX-style symbolic links
    (e.g.: Windows!);

 2. We cannot leave even a symbolic link behind or the Tizen SDK's
    packaging utility will bundle two copies of the library.

For these reasons, we copy libmonoboehm-2.0.so.1 to libmonoboehm-2.0.so
just-in-time before shelling out to native-make, and remove it immediately
after.