csproj: Trim list of directly referenced assemblies
This is an attempt to work around this error:

    /usr/lib/mono/4.0/Microsoft.Common.targets:  warning : Reference
    'glade-sharp, Version=2.12.0.0, Culture=neutral,
    PublicKeyToken=35e10195dab3c99f' not resolved

occurring under the addins.monodevelop.com autobuilder.  While we
don't reference `glade-sharp`, it looks like the list can be trimmed a
bit.

One of the files is explicitely "using Gtk;", and inheriting from
Gtk.Dialog won't work without a reference to atk-sharp.